updates to be more faithful to original
