+++ src/index.ts
@@ -1,4 +1,4 @@
-export { CentralStateMachine } from './stateMachineCentral'
+export { BotStateMachine } from './stateMachineBot'
 export { NestedStateMachine } from './stateMachineNested'
 export { StateBehavior, StateTransition } from './stateBehavior'
 export { StateMachineWebserver, WebserverBehaviorPositions } from './webserver'
